backend/pyproject.toml
[tool.poetry]
name = "helios-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^=3.0.2"
daphne = "^2.4.1"
django-environ = "^=0.4.5"
psycopg2-binary = "^=2.8.4"
djangorestframework = "^=3.11"
markdown = "^=3.1.1"  # for DRF browsable APIs
django-filter = "^=2.2.0"  # for dynamic filtering of querysets
pydantic = "^1.4"
django-rest-auth = {extras = ["with_social"], version = "^0.9.5"}
django-allauth = "^0.41.0"


[tool.poetry.dev-dependencies]
django-extensions = "^=2.2.5"
django-debug-toolbar = "^=2.0"
Werkzeug = "~=0.16.0"
factory_boy = "~=2.12.0"
pytest = "~=5.3.2"
coverage = "~=5.0.3"
factory-boy = "~=2.12.0"
pytest-cov = "~=2.8.1"
pytest-django = "~=3.5.1"
pytest-xdist = "~=1.30.0"
pytest-sugar = "~=0.9.2"
pytest-faker = "~=2.0.0"
pytest-mock = "~=1.11.1"
pytest-pudb = "~=0.7.0"
pytest-factoryboy = "~=2.0.3"
black = "^19.10b0"
django-stubs = "^1.4.0"
ipython = "^7.13.0"
faker = "^4.0.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
